Implement script to generate olm configuration JSON file for keyserver
Summary: Implemented script that uses olm library to create and pickle data (encryption keys) needed to establish sessions with keyserver as a client. It was discussed during a meeting that we do not expect pickling key to provide additional security layer so I simply used standard uuid library to generate it.

Test Plan: Run script both when the file is present to see it is overwritten and when it is missing to see it is created.

Reviewers: palys-swm, varun, def-au1t, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: benschac, ashoat, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3209