[web] Introduce empty account page
Summary:
An empty page with only a header. This page is not rendered yet, so there are no visible changes in the app. The rendering will be handled in much later diff.
Designs can be found here: https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=1272%3A88750

{F17253}

Test Plan: Display the page instead of e.g. Calendar and check if it is displayed correctly

Reviewers: benschac, atul, def-au1t

Reviewed By: benschac, atul

Subscribers: ashoat, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3320