Move SQLiteQueryExecutor initialization to common place and implement it in a thread-safe way
Summary: If we wish to access database from different places on mobile devices, possibly when JS is not started we need a common function to call, that is also idempotent.

Test Plan: Build both android and iOS app and examin logs after switching between background and foreground several times to ensure that SQLiteQueryExecutor globals are initialized exactly once.

Reviewers: palys-swm, atul, karol-bisztyga, ashoat

Reviewed By: palys-swm, karol-bisztyga, ashoat

Subscribers: benschac, ashoat, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3296