[web] Remove `background.png` from `Splash` and rename to `placeholder.png`
Summary:
Here's how it looks: {F23212}

---

depends on D3445

Test Plan:
Did a search for "images/background.png" and found no other occurences.

Then did a search for "background.png" and found that we were actually using it in `sanitization.js` to replace image URLs as part of the "sanitization process." So I renamed `background.png` to `placeholder.png` and left the file in.

Reviewers: def-au1t, palys-swm, varun, benschac, ashoat

Reviewed By: ashoat

Subscribers: Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3446