[web] Delete extraneous code from `PasswordChangeModal`
Summary:
Delete all the code that is not needed in the modal and most of the unused css. The code will be simplified and modified (e.g. change copy) in the next diff. The css will be updated in some later diff, when the design is updated.

{F24944}
Note: the purpose of this diff is to delete components from the modal. Updating its design is done later.

Depends on D3472

Test Plan: Display a modal and check if it contains only relevant components. Also check if error messages are displayed when needed and if password change functionality works.

Reviewers: benschac, atul, def-au1t

Reviewed By: benschac, atul

Subscribers: ashoat, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3473