[web] [fix] style mark as unread
Summary:
update mark as read. https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=1085%3A85147

We decided to switch the background color for hover state mark as unread button.

{F24737}

to option one.
the selected chat thread item (mochi) is the selected state which is not changing from the design.
the reason we needed to update the background color for the hover state mark as unread button is because the background color of hover chat thread item and background color of mark as unread button are the same color in the figma file.

Test Plan: check figma. We still have similar issues around the hover state leaving too soon on mouse leave which need to be fixed in another diff. This is just updating the styling to match our spec.

Reviewers: atul, palys-swm, def-au1t, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3422