[web] Introduce search modal component
Summary:
We're going to have a couple of modals that have search functionality, so it makes sense to have a component for that.
We usually prefer using children components when introducing containers like this, but in this case that approach has a couple of disadvantages: we would need to pass `searchText` to children as a prop and that would require using `React.cloneElement`. Also, defining child component props would be complicated to allow supporting all the components that have `searchText` props. None of these are too serious, but it looks like using children function is more readable and easier approach. If you disagree, it can easily be modified.
Depends on D3385

Test Plan: Open members modal and check if filtering works correctly.

Reviewers: benschac, atul, def-au1t

Reviewed By: benschac, atul, def-au1t

Subscribers: ashoat, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3494