[web] Introduce common `Search` component
Summary:
As we have design for search for filtering users in modals, the goal is to use the same component for both thread search (for which we don't have design) and thread members search.
This diff moves search logic and styles from `thread-list-search` into `search`.
There are no design or behavior changes in this diff - everything should work as before.

Test Plan: Thread search in web app should remain the same

Reviewers: palys-swm, benschac, atul

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D3367