[web] [feat] ENG-766 add arrow extender to thread sidebar
Summary:
adds an extender to remove gap between lines. It's not perfect. I need to figure out a way to make it blur so it matches the svg more closely. If you zoom in you can see it but, it's very very slight. This diff will unblock the rest of the stack

before: {F26747}
after: {F26751}

Test Plan: add many threads to a conversation, arrows should connect. https://linear.app/comm/issue/ENG-941/arrow-extenders-should-match-blur-of-svg

Reviewers: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3539