[services][identity] initial commit for the identity service.
Summary: `cargo new` does most of the scaffolding for us. i made a small change to the .gitignore file to make sure we don't check in targets. we do, however, want to check in the Cargo.lock file since this is a binary.

Test Plan: `cargo build && cargo run` works.

Reviewers: jimpo, karol-bisztyga, palys-swm, geekbrother

Reviewed By: jimpo, karol-bisztyga, palys-swm, geekbrother

Subscribers: geekbrother, ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3569