[services][identity] Dockerfile for the identity service
Summary:
We want to containerize the identity service so we can run it on any system that has a Docker engine. This Dockerfile:
1. Uses Rust as the base image
2. Creates a new empty shell project
3. Copies over the manifests (Cago.toml and Cargo.lock)
4. Builds the app's dependencies to cache them
5. Copies over the source code
6. Builds a release version of the binary
7. Runs the binary

Depends on D3590

Test Plan: Built the docker image with `docker build -t identity .`. Then ran the container with `docker run -dp 50054:50054 --rm --name identity1 identity`. Then used BloomRPC to make calls to the service.

Reviewers: jimpo, karol-bisztyga, atul

Reviewed By: jimpo

Subscribers: ashoat, palys-swm, Adrian, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3593