[services] Backup - Add hash to send log request
Summary:
Depends on D3561

I think for every operation when we possibly will update something to the blob service, we need to require a hash of the data that's about to be uploaded. This is because in the blob service we require a hash for the `Put` operation.

Since we decided to store large logs in the blob service, we have to also require a hash here.

Test Plan:
```
cd services
yarn run-backup-service
```

Reviewers: palys-swm, geekbrother, jimpo, varun, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3562