[web] Link wordmark in header to app "homepage"
Summary:
Addressed issue where the "Comm" wordmark in header did not link to the app "homepage". Now, when a user hovers over the "Comm" wordmark on a page "Comm Home" appears and clicking on the wordmark will link to the app root/homepage.

Video:
{F28639}

More context on [[ https://linear.app/comm/issue/ENG-953/link-wordmark-in-header-to-app-homepage | Linear ]].

Test Plan: Ensured that the Comm wordmark on different pages on the web app linked back to the root page with all the chats. Tested on Chrome/Safari.

Reviewers: atul, ashoat, benschac

Reviewed By: atul, ashoat, benschac

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac

Differential Revision: https://phabricator.ashoat.com/D3604