[services] Dev Mode - create backups for AWS S3 and dynamoDB
Summary:
I think it would be good to keep track of the dumps from dynamo DB tables and the S3 buckets.
Every time someone changes a table, they should run the backup script, so the values get updated in our repository.
We can then easily recreate the S3 bucket structure and all the tables (without data) from the dynamo DB.
This can be useful for testing as well as for the tasks like setting up a local cloud (this is done in the upcoming diffs).
In general, I think having a backup of the structure of these functionalities like this is beneficial, what do you think?

Test Plan:
```
cd services
yarn backup-aws
```

Reviewers: jimpo, palys-swm, geekbrother, varun

Subscribers: ashoat, palys-swm, Adrian, atul, benschac

Differential Revision: https://phabricator.ashoat.com/D3495