[docs] Use a double ampersand in chained bash commands
Summary: Instead of running the second (right) bash install command anyway (even if the first one failed), we should use `&&` to run it only when the first (left) command is successful. This is solved by switching from `;` to `&&` when chaining the commands together.

Test Plan:
For example: running `brew install php@7.4; brew upgrade php@7.4` the first (left) command can fail in case the current version is unavailable or due to the local brew dependencies problem. In that case, the right command will be invoked anyway and of course, will fail too.
If we chain the commands with the `&&` instead, the execution will stop at the first failed command and we will not execute the next one in the commands chain.

Reviewers: varun, ashoat

Reviewed By: varun, ashoat

Subscribers: varun, ashoat, palys-swm, Adrian, atul, karol-bisztyga, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3685