[web] Fix "See more..." button alignment in `ChatThreadListSeeMoreSidebars` component
Summary:
Addressed issue where the "See more..." button wasn't aligned with the rest of the items in the `ChatThreadList`.

I saw that Ben and Ashoat mentioned the height of the sidebar item changing in the Linear issue. I'll defer to a later diff to refactor the styling of the "See more..." button if the height change of the sidebar affects anything there.

Before:
{F26786}

After:
{F26787}

More context: https://linear.app/comm/issue/ENG-920/see-more-threads-button-doesnt-look-right-on-web

Test Plan: Ensured that thread list looked correct visually. Tested in Chrome/Safari at different breakpoints.

Reviewers: atul, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac

Differential Revision: https://phabricator.ashoat.com/D3541