[web] Introduce minimally viable `ThreadSettingsGeneralTab` component
Summary:
In an effort to make `ThreadSettingsModal` easier to work with, we're pulling the `general`, `privacy`, and `delete` tabs out into their own separate components.

This diff does that for `ThreadSettingsGeneralTab`, but in a limited way. We're basically moving over the `mainContent` JSX markup and prop drilling the state that's still held by `ThreadSettingsModal`.

Once we've extracted the JSX markup for `general`, `privacy`, and `delete`.. and prop drilled as necessary to keep things working as they did before: we're going to push down the state and functionality into the smaller "tab" components so things are properly encapsulated.

And once that's done, we should easily be able to turn `ThreadSettingsModal` into a functional component.. and refactor `ThreadSettingsModal` using @def-au1t's `TabsContainer` and `TabsItem` components.

Test Plan:
This diff should be a noop. I spent some time changing the thread title, description, and color--and things continue to look and work as expected:

{F38144}

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga, benschac, yayabosh, def-au1t

Differential Revision: https://phabricator.ashoat.com/D3716