[web] Introduce minimally viable `ThreadSettingsPrivacyTab` component
Summary:
In an effort to make `ThreadSettingsModal` easier to work with, we're pulling `general`, `privacy`, and `delete` tabs out into their own separate components.

This diff does that for `ThreadSettingsPrivacyTab`, but in a limited way. We're basically moving over the `mainContent` JSX markup and prop drilling that state that's still held by `ThreadSettingsModal`.

Once we've extracted the JSX markup for `general`, `privacy`, and `delete`.. and prop drilled as necessary to keep things working as they did before: we're going to push down the state and functionality into the smaller "tab" components so things are properly encapsulated.

And once that's done, we should easily be able to turn `ThreadSettingsModal` into a functional component.. and refactor `ThreadSettingsModal` using @def-au1t's `TabsContainer` and `TabsItem` components.

---

Depends on D3716

Test Plan:
This diff should be a noop. I spend some time changing the thread privacy settings--and things continue to look and work as expected (we'll likely need a restyle of this tab, but that's for another diff):

{F38182}

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga, benschac, yayabosh, def-au1t

Differential Revision: https://phabricator.ashoat.com/D3717