[services] Backup/Blob - fix handling reading done in server reactors
Summary:
Depends on D3562

we should suppress this as we want to have the ability to gracefully end a connection on the other side. This will result in `!ok` here. I think it is somehow broken and the simple bool flag doesn't give us enough information on what happened.

We should manually check if the data we received is valid

Test Plan: The same as in for instance D3535, we have to see that the connection end on the other side(so if we're looking at the server, the connection should end gracefully at the client and vice versa).

Reviewers: geekbrother, varun, palys-swm, jimpo

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3582