[web] Use relationship callback on friend list
Summary:
Handle `REQUEST_SENT` and `REQUEST_RECEIVED` states - call callbacks when user clicks the buttons.
Depends on D3735
Depends on D3677

Test Plan:
Send a request from native and check if it is possible to cancel it from web.
Receive a request and check if it is possible to accept or reject it.

Reviewers: benschac, atul, def-au1t

Reviewed By: atul

Subscribers: ashoat, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3736