[lib] [feat] [ENG-991] add usePromoteThread hook
Summary: abstract promote thread into a hook for use in both native and web. We're using the same API call and reducer in web, all we need to do is pull the call in native into lib so we can use it in two places.

Test Plan: n/a hook is not used yet, code is a copy/paste from: https://github.com/CommE2E/comm/blob/master/native/chat/settings/thread-settings-promote-sidebar.react.js in native.

Reviewers: atul, palys-swm, def-au1t, yayabosh, ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3746