[web] Move `currentTabType` state up from `ThreadSettingsModal` to `ConnectedThreadSettingsModal`
Summary:
Move `currentTabType` state out of the inner class component state, and move it out to the wrapping "Connected" functional component.

As part of the work to refactor `ThreadSettingsModal`… specifically to turn it into a functional component.

---

Depends on D3757

Test Plan:
Close reading... should be a noop

Will test the entirety of `ThreadSettingsModal` thoroughly once the refactor is complete (plan on using this as an opportunity to write some automated tests)

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3759