[native] Introduce `ColorSelectorButton` component
Summary:
First step in redesigning the `ColorPicker` component on `native` to match the `ColorSelector` component on `web` introduced by @atul.

`ColorSelectorButton` is a new component that will work with `ColorSelector` to replace `ColorPicker` on `native`.

See the full plan for the `ColorPicker` redesign in this [[ https://linear.app/comm/issue/ENG-1001/redesign-native-colorpicker-component | Linear issue ]].

Test Plan:
Tested on iOS 15.4 and works as expected:
{F42296}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, benschac, atul

Differential Revision: https://phabricator.ashoat.com/D3790