[native] Introduce `ColorSelector` component
Summary:
Part of the first step in redesigning the `ColorPicker` component on `native` to match the `ColorSelector` component on `web` introduced by @atul. Depends on D3790

`ColorSelector` is a new component that will work with `ColorSelectorButton` to replace `ColorPicker` on `native`. This diff handles color layout, and adding `onPress` handling and active state functionality will be applied in later diffs.

See the full plan for the ColorPicker redesign in this [[ https://linear.app/comm/issue/ENG-1001/redesign-native-colorpicker-component | Linear issue ]].

Test Plan:
Tested on iOS 15.4 and works as expected:
{F42549}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian, karol-bisztyga, atul, benschac

Differential Revision: https://phabricator.ashoat.com/D3798