[services] Backup - Add data exists flag to put client reactor
Summary:
Depends on D3774

We need to keep track of whether the data already exists on the cloud or not. We need this, because, in a case, it is already there, we want to just add a new holder to it without uploading the same bytes.

https://linear.app/comm/issue/ENG-672/create-addreferencetohashbyholder-in-the-blob-service

Test Plan:
```
cd services
yarn run-backup-service
```

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3775