[services] Backup - synchronize status access in reactors
Summary:
Depends on D3789

The issue was raised in https://phabricator.ashoat.com/D3720#103334 and I decided it would be the best to handle synchronization in the mutual code, therefore I first wanted to implement pulling services' mutual code into a single master class.

This is for the backup service.

Test Plan:
```
cd services
yarn run-backup-service
```

Reviewers: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3800