[keyserver] Add nvm to Dockerfile
Summary:
Our build/run scripts for the keyserver rely on `nvm`, so we need it in the `Dockerfile`.

Trying to lock to a specific version here. Inspired by [this StackOverflow](https://stackoverflow.com/questions/25899912/how-to-install-nvm-in-docker) and [this article](https://tecadmin.net/how-to-install-nvm-on-debian-11/).

Depends on D3848

Test Plan: tested that `yarn prod` worked inside the Docker image

Reviewers: palys-swm, jimpo

Reviewed By: palys-swm

Subscribers: Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3849