[web, keyserver] [fix] [ENG-1022] add svg images to sidebar to try and fix arrow issue
Summary:
This is a diff to show what's going on when I try to just add an svg as an image. Attaching a screenshot. But, it's looking like the extended arrow isn't the same width, arrow thickness, or shape as the original arrow. I tried updating both the svg view box and height + width properties without much luck.

{F45278}

Test Plan: N/A, this diff is just to document and show the team what the "use an image" html element path looks like.

Reviewers: atul, palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3841