[web] Lift `tabs: React.Node` from `ThreadSettingsModal` to `ConnectedThreadSettingsModal`
Summary:
Lift `tabs` from the `ThreadSettingsModal` class component to the `ConnectedThreadSettingsModal` functional component.

Doing this to "stage" the lifting of JSX markup into separate steps to hopefully make it easier to review.

As part of the work to refactor `ThreadSettingsModal`... specifically to turn it into a functional component.

---

Depends on D3823

Test Plan:
1. Open `ThreadSettingsModal`
2. Make sure the "tabs" of the `ThreadSettingsModal` component continue to look and behave as expected (since that was what was moved)
3. Change some settings around, hit save, and make sure things are saved as expected

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3824