[web] Handle longer usernames in friend list
Summary:
When a username is longer than available space, we should display an ellipsis at its end, just like we do in other places. We also need to make sure that the text in a button doesn't wrap.

{F46656}

Test Plan: Send a friend request to a user with long username and check if the name is displayed correctly on friend list.

Reviewers: benschac, atul, def-au1t

Reviewed By: atul

Subscribers: ashoat, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3862