[web] Introduce `AddMembersModal` component
Summary:
Introduce "AddMembersModal", that is used for searching and selecting users to add to a thread.
There is no logic for searching users, and for dispatching server action, added in this diff yet - they will be added in the following diffs.

https://linear.app/comm/issue/ENG-966/allow-adding-users-to-the-thread

Demo:
{F39829}

Test Plan: The modal will be displayed after pressing "Add members" button in MembersModal - after introducing the action in the following diffs.

Reviewers: palys-swm, atul, benschac

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3742