[web] Introduce searching users in `AddMembersModal`
Summary:
Add logic responsible for finding users, that can be added (or not) to the thread based on search query. Used `getPotentialMemberItems` lib function, that is used in native app.

{F39830}

Test Plan: The search can be tested after introducing button activating `AddMembersModal` - search results should be exactly the same as in mobile app for the same query.

Reviewers: palys-swm, atul, benschac

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, benschac, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3743