[web] Display unfriend action menu
Summary:
When an edit button is clicked we display a menu where a user can unfriend another user

{F54382}

https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=1311%3A94083

This diff is affected by https://linear.app/comm/issue/ENG-1125/opening-new-menu-without-closing-the-previous-one-doesnt-work which also affects the master.

Depends on D4006

Test Plan:
Click an edit button and check if menu is displayed.
Click outside of it and check if it closes.
Open it and click another edit button - verify that only one menu is opened at a time.
Open menu and check if `unfriend` action works.

Reviewers: benschac, atul, def-au1t

Reviewed By: benschac, atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4007