[services] Tunnelbroker - Switch to milliseconds at `createdAt` in messages
Summary:
Switch to milliseconds from unixtime seconds for the `createdAt` field on Tunnelbroker database message item related to the
[[ https://linear.app/comm/issue/ENG-1027#comment-0ef7fada | ENG-1027 comment ]].

Using the `tools::getCurrentTimestamp()` to get the milliseconds since epoch.

Related task: [[ https://linear.app/comm/issue/ENG-1045/switch-to-milliseconds-from-seconds-in-tunnelbroker-message-time-of | ENG-1045 ]]

Test Plan: Run `yarn test-tunnelbroker-service` to build and run the tests.

Reviewers: karol-bisztyga, palys-swm, varun, ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3880