[keyserver] Update Node Docker Compose config to always restart
Summary:
MySQL takes a while to start up, and Node will crash if MySQL (or Redis) doesn't respond.

We should probably make it so Node doesn't crash if it can't reach a service, and instead keeps retrying. But I actually think it's good to auto-restart the Node server either way here.

Test Plan: Run `docker-compose up --build` and make sure Node keeps restarting when it fails to connect to MySQL

Reviewers: jimpo, varun

Reviewed By: varun

Subscribers: palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3884