[web] Add background illustration to background tab
Summary:
We already did this on mobile a while ago.

As part of https://phabricator.ashoat.com/D3856 ("[native] Bump `react-native-svg` to `12.3.0`") I double checked that the background illustration continued to look as expected on `native`... at which point I realized we never made the corresponding change on `web`... so decided to do it really quick.

Test Plan:
Here's what it looks like:

{F49064}

Reviewers: def-au1t, palys-swm, benschac, ashoat

Reviewed By: ashoat

Subscribers: Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3889