[web] Remove `max-height` property from `modal.css: modalContainer[Small/Large]`
Summary:
Looks like this property was introduced in D3669. I understand fixing the width of the modals, but I think we can let the height expand to fit the content?

It doesn't seem right to arbitrarily cap the possible height of a Modal at this level of abstraction? I understand there might be the need to cap it for specific Modals?

For example, here's the `ThreadSettingsModal` before and after this diff:

Before:
{F49120}

After (and before D3669):
{F49119}

Test Plan:
Go through every usage of `Modal` and make sure things looks as they did before.

(Haven't actually done this yet, but will do before landing)

Reviewers: def-au1t, palys-swm, benschac, ashoat

Reviewed By: def-au1t, ashoat

Subscribers: Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3892