[web] [chore] [ENG-530] remove old styles from inline sidebar component
Summary: remove all old styles and icons since the new design is pretty different from what we have. This diff shouldn't be landed on it's own. More just clean up for updating the component

Test Plan: N/A additionially, there's literally nothing to see here. The text is black by default so you're not going to see anything on the screen.

Reviewers: atul, ashoat

Subscribers: palys-swm, Adrian, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3893