[services] Add depends_on
Summary:
Adding `depends_on` to the services so they depend on the localstack.

This was requested in D3692.

Please, note that if we decide that we may want to connect to the external cloud service from the local environment, then this change will have to be reverted.

Related task https://linear.app/comm/issue/ENG-1056/discuss-dev-mode-design

Test Plan:
```
cd services
docker-compose down
yarn run-backup-service
```
This will create localstack container on the way automatically

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3958