[services] Tunnelbroker - RabbitMQ Docker image for dev environment
Summary:
For a local Tunnelbroker development environment, we need to use a local instance of the RabbitMQ server.
Add the RabbitMQ docker image after the localstack image to the `docker-compose`.
AMQP(s) ports 5672 and 5671 are forwarded to use in local or docker builds.
The rabbitMQ management console is listening on `localhost:15672`.
The local image user and password are set to `comm:comm`.

Test Plan: Successfully connected to the rabbitMQ by AMQP client and logged into the local management web-gui.

Reviewers: karol-bisztyga, palys-swm, varun, jimpo, ashoat

Reviewed By: karol-bisztyga, palys-swm, jimpo, ashoat

Subscribers: jimpo, ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D3964