[web] Introduce block list row
Summary:
Create a component that is displayed as a row on block list. The design can be found here https://www.figma.com/file/a1nkbWgbgjRlrOY9LVurTz/Comm-%2F-Desktop-app?node-id=1313%3A98810.

{F55677}

Depends on D4036

Test Plan: Check if all blocked users are displayed. Unblock a user and check if the row is no longer present on the list.

Reviewers: benschac, atul, def-au1t

Reviewed By: atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4037