[keyserver] Use importJSON to import olm_config.json
Summary:
It seems that this function is not used anywhere, so I am hoping it's okay to switch to an async function.

(The reason for this is that I want to use `importJSON` so I can later refactor all configs so they can be specified by environmental variables in the Docker context.)

Depends on D4024

Test Plan: Call `getOlmConfig` at the end of `generateOlmConfig` and make sure it works. Diff: https://gist.github.com/Ashoat/3b4ad12b59c54847f8f277914d12837e

Reviewers: marcinwasowicz, palys-swm

Reviewed By: marcinwasowicz, palys-swm

Subscribers: Adrian, atul, marcinwasowicz, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4025