[web] Clean up `NotificationsModal` SVG assets and remove `SWMansionIcon` dependency
Summary:
Linear issue: https://linear.app/comm/issue/ENG-1141/use-svg-instead-of-swmansionicon-in-notificationsmodal

We were previously including the SVG illustrations in `NotificationsModal` via `SWMansionIcon` which requires going through the whole "IcoMoon" process. As noted in the Linear issue, the illustrations didn't look great (especially on Safari).

Exported the icons from Figma as SVG, cleaned up the vectors in Sketch, re-exported the cleaned-up illustrations as SVG, uploaded to S3, and replaced `SWMansionIcon` with `<img>` in `NotificationsModal`.

Before:
{F56220}

After:
{F56221}

Test Plan: Ensured that the illustrations looked as expected on Safari, Chrome, and Firefox.

Reviewers: def-au1t, palys-swm, benschac

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4041