[web] Rename `NotificationsOption` -> `EnumSettingsOption`
Summary:
We want to reuse the (previously known as) `NotificationsOption` component for the "Privacy" tab of `ThreadSettingsModal`.

As a first step, we rename `NotificationsOption` to more generic `EnumSettingsOption` (open to alternative names). In later diffs will remove "Notifications-specific" parts of `EnumSettingsOption` and make it more generic.

Made sure to rename component, rename file, and fix imports.

---

Depends on D4044

Test Plan: Opened `NotificationsModal` and it continues to look/work as expected. Also flow/eslint.

Reviewers: def-au1t, palys-swm, benschac

Reviewed By: palys-swm, benschac

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4045