[native, lib, web] [refactor] move text out of state
Summary: to make search sidebars work with the SearchModal component it's state needs to be refactored to only the result set. This diff removes the text key/value and moves it into a React.useState which is later refactored in the following diff to get sidebar search data to work with SearchModal component

Test Plan: Flow and eslint don't error. The sidebar search isn't functional in this diff but will be in the next diff.

Reviewers: @atul, @def-au1t, @palys-swm

Subscribers: