[web] Fix thread description and color modification in private thread
Summary:
Fix issue where user is unable to modify the thread description and color in their private thread. The thread name is still unmodifiable (following the inclusion of the `EDIT_DESCRIPTION` and `EDIT_COLOR` permissions but not the `EDIT_THREAD_NAME` permission in D1980), but the thread description and color can be modified.

Related Linear issue with full context [[ https://linear.app/comm/issue/ENG-1046/cannot-change-thread-color-in-private-thread | here ]].

Test Plan:
Tested on Chrome/Safari and works as expected:

Before:
{F59147}

After:
{F59150}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D4073