[web] Introduce danger zone route
Summary:
Add a new route for danger zone. It looks like this is the first time we introduce a route consisting of two words, but using hyphens sounds like a better approach compared to underscores.

{F59578}

Depends on D4091

Test Plan: Open settings and check if two nav items are displayed. Try switching between them - always the active should be highlighted. When danger zone is opened, an empty page should be displayed.

Reviewers: def-au1t, atul, benschac, ashoat

Reviewed By: atul, ashoat

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4093