[web] Use `EnumSettingOption` components in `ThreadSettingsPrivacyTab`
Summary:
This is a minimally-viable first pass.

Replace the inputs in `ThreadSettingsPrivacyTab` with `EnumSettingOption`s. There is still a good bit to go with stying, but should be functionally complete.

Here's how it looks:
{F56474}

---

Depends on D4050

Test Plan: Click around, change visibility, observe that things look/behave as expected

Reviewers: palys-swm, def-au1t, benschac

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4051