[web] Introduce account delete modal
Summary:
This modal contains a part of the logic from user settings modal, so it is a good idea to reuse what we already have there. The first step is to simply copy the content of that modal (with naming and import changes) - we can do that, because the last step of this stack will be to delete the original modal which will get rid of the duplication.

This is the first step in adding the modal, introduced mainly to allow easier review of the code - so that it is clear which parts stay and which are not needed.

{F61769}

Test Plan: Modify `danger-zone` so that the modal is displayed after user clicks `Delete Account` button. A modal that looks like user settings should be displayed.

Reviewers: def-au1t, atul, benschac

Reviewed By: atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4117