[keyserver] Extract getJSON from importJSON
Summary:
This diff is a pure refactor. In the next diff I extend `getJSON` a bit, and it helps readability a little bit to extract it into its own function beforehand.

Depends on D4098

Test Plan: I tested this in combination with the next diffs, which makes sure `importJSON` still works correctly

Reviewers: palys-swm, atul

Reviewed By: atul

Subscribers: Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4101