[web] Hide thread menu for pending threads
Summary:
The diff solves the issue:
https://linear.app/comm/issue/ENG-1168/thread-menu-is-available-for-pending-threads-and-crashes-the-app
Is prevents rendering thread menu (the one on the right of top thread bar) when thread is pending.

Test Plan: Create pending sidebar (or thread by writing to the new user). The menu is not displayed now, and it's not possible to crash the app by launching any thread settings modal)

Reviewers: palys-swm, atul, benschac

Reviewed By: atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4092