[web] Increase default size of thread description `textarea` in `ThreadSettingsGeneralTab`
Summary:
Increased the default size of the `textarea` that contains the thread description in `ThreadSettingsGeneralTab` to display 3 rows. This is because the old size did not fully display the default thread description ("This is your private thread, where you can set reminders and jot notes in private!") with the last line cut off.

Related Linear issue with full context [[ https://linear.app/comm/issue/ENG-1161/increase-default-size-of-thread-settings-description-input-to-fit | here ]].

Test Plan:
Tested on Chrome/Safari and works as expected.

Before:
{F61044}

After:
{F61045}

Reviewers: atul

Reviewed By: atul

Subscribers: ashoat, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D4104