change ios notification structure to make it visible for NotificationService
Summary: If we want the system to launch NotificationService when notification is received it muse have flag mutable-content set to true (https://developer.apple.com/documentation/usernotifications/unnotificationserviceextension?language=objc)

Test Plan: This change was already used to test previous differentials in this stack. The only additional testing needed is to check whether it doesn't break previous versions of the app (no NotificationService).

Reviewers: tomek, ashoat

Reviewed By: tomek, ashoat

Subscribers: ashoat, adrian, atul, abosh

Differential Revision: https://phab.comm.dev/D4113