[native][lib] "Permanently" make changes for release
Summary:
We made changes to `file-utils:mediaConfig` and `media-gallery-keyboard:fetchPhotos` to accommodate video message work while it was in progress. The change in `file-utils` was for treating animated GIFs as videos... and the change in `media-gallery-keyboard` was for including videos in the media selection gallery.

Since the ability to send video messages still isn't ready, the changes have to be reverted before every release... and "un-reverted" after each release. This gets pretty tedious and adds some overhead to the release process (two extra commits).

Since video messages aren't actively being worked on, we decided in ENG-170 (https://linear.app/comm/issue/ENG-170/daily-iosandroid-releases) that it was reasonable to make the changes "permanently."

I left some comments in the code near the changes with steps to re-introduce them when work on video messages starts back up.

Test Plan: NA

Reviewers: def-au1t, palys-swm, ashoat

Reviewed By: ashoat

Subscribers: Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4123