[services] Backup - Create new backup - avoid suppressing database errors
Summary:
Depends on D4077

When creating a new backup I mistakenly just printed out database errors to stdout on the server. That led to an unwanted situation when something went wrong with the database and the client received a positive outcome. With this change, the error breaks uploading a new backup and the client receives proper information on what went wrong.

Test Plan:
Services can still be built.

It's hard to really test this. In order to do this, we'd have to intentionally hardcode invalid instructions on the database and then see the result on the client-side.

You can do that, please see D4074 on how to test the services with the client.

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4078