[services] Tunnelbroker - Add yarn scripts for dev-mode
Summary:
To run the tunnelbroker in dev-mode we need to export `COMM_SERVICES_DEV_MODE` environment variable.
Following the current code approach, add the corresponding commands for the tunnelbroker to `package.json`.

Linear task: [[ https://linear.app/comm/issue/ENG-1101/read-env-variables-to-change-the-dynamodb-table-names-and-amqp-uri-in | ENG-1101 ]]

Test Plan: Run `run-tunnelbroker-service-dev-mode` and check that the `tools::isDevMode()` is true.

Reviewers: karol-bisztyga, palys-swm, ashoat

Reviewed By: karol-bisztyga, palys-swm, ashoat

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4126