[services] Tunnelbroker - Dev-mode environment variables in Docker
Summary:
To use the dev-mode inside the tunnelbroker we need to pass the `COMM_SERVICES_DEV_MODE` environment variable to the Docker container itself.

Linear task: [[ https://linear.app/comm/issue/ENG-1101/read-env-variables-to-change-the-dynamodb-table-names-and-amqp-uri-in | ENG-1101 ]]

Test Plan: Run `yarn run-tunnelbroker-service-dev-mode` shows that the environment variable `COMM_SERVICES_DEV_MODE` is equal to `1`.

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, karol-bisztyga, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4131