[services] Backup - Add Attachment Reactor Logic
Summary:
Depends on D4134

Adding `AddAttachments` logic. This is a Unary Reactor which means it goes only for one cycle, the response is empty, so it effectively just receives the data on the server and then returns a status.

`logID` is optional. If it's an empty string, the attachments are appended to the backup item.

Test Plan:
- terminal 1
```
cd services
yarn run-blob-service-dev-mode
```
- terminal 2
```
cd services
yarn run-backup-service-dev-mode
```
- terminal 3
```
git clone https://github.com/karol-bisztyga/grpc-playground.git
cd grpc-playground
git checkout backup-async
./build.sh
./cmake/build/bin/client
```
- in terminal 3 create a new backup with `n`
- check the database and see the empty string in `attachmentHolders` in the newly created item
- add new attachments: in terminal 3 use `b`
- check the database again and see new attachments in `attachmentHolders` in the newly created item
- do a similar thing for the logs: first, use `l` and then `k`

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4135