[keyserver] Extract migration config to separate file
Summary:
In the following diff, I'm going to check the latest migration index from `setup-db.js` so we can start with a default migration version.

Without this diff, that change would result in a require cycle.

Depends on D4151

Test Plan: This is a simple refactor. I tested it in combination with stack (see test plan in D4151), and also used Flow

Reviewers: palys-swm, atul

Reviewed By: palys-swm, atul

Subscribers: Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4152