[native] Add `messageStoreLoaded` to `SQLiteContext`
Summary:
We use `SQLiteContext` to determine whether the state persisted in SQLite has been loaded.

For example, in `app-navigator`, we won't render within `App.Navigator` until `storeLoadedFromLocalDatabase` is true.

---

Depends on D4158

Test Plan:
We're manually setting `messageStoreLoaded` to true in `SQLiteContextProvider` for `persistConfig.version` < 31 (the migration which "flips the switch") for now.

The app should continue to work as expected without issue. Specifically, will deploy a release build to my phone and open/force quit/open/etc the app a couple times and make sure things continue to work as expected.

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4159