[lib] Exclusively use "ops approach" in `reduceThreadUpdates`
Summary:
`reduceThreadUpdates` used to return both an updated `threadStore` object and corresponding `threadStoreOperations`. `reduceThreadUpdates` has a single callsite and that's in `reduceThreadStore` where only the `threadStoreOperations` property of the returned object was being pulled out anyways... so in this diff we remove the updated object to clean things up.

(Since this function is no longer really a "reducer," we'll rename this to something like `generateOpsForThreadUpdates`)

---

Depends on D4196

Test Plan: `assertThreadStoreThreadsAreEqual` hasn't flagged anything since it's been included in staff builds so we can assume "the ops are good."

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4197