[lib] Update `messageStore.messages` with `msgStoreOps` for logIn.success action
Summary:
Instead of getting `messageStore.messages` from the `freshStore` returned by `freshMessageStore`... we use the `messageStoreOps` returned by `freshMessageStore` to compute the new message store.

Note that a `messageStore.messages` is still being created within `freshMessageStore(...)` and will be removed in a subsequent diff

Test Plan: This should be a straightforward refactor given we've had `assertMessageStoreMessagesAreEqual` in place for some time now.

Reviewers: def-au1t, palys-swm

Reviewed By: palys-swm

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4201