[keyserver] Make it possible to run Comm web app without Apache proxying to it
Summary:
To avoid breaking anybody's environment, if the `proxy` config is left out, we'll continue defaulting to Apache.

This diff makes it easier to test things sometimes, and perhaps somebody will want to run the Web app without a proxy. (That said, we don't currently support HTTPS/TLS without a proxy, so it would probably be a bad idea to do that on production.)

Linear task: https://linear.app/comm/issue/ENG-1221/make-it-possible-to-run-node-without-apache

Test Plan: I tested a local environment proxied over Apache and a local environment without proxying. I did not test a local environment without proxying with HTTPS enabled, as this would require adding support for certificates to the Node server, which is out of scope

Reviewers: atul, palys-swm

Reviewed By: palys-swm

Subscribers: Adrian, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4210