[services] Identity - Fix proto file location
Summary:
I think we should have all proto files in one place. Therefore, here I'm moving the proto for identity service to `native` where all protos are. We can either do that or move all protos to another location but I'd have them in one place.
Additionally, I changed the folder name on the docker image from `proto` to `protos`. I this plural is better here because we may have multiple protos in each service (in the case of interservice communication, we already have that in the backup service), plus it's like this in other services so we should probably stay consistent about this.

Test Plan:
Not sure how Varun has been testing the identity so far but this should work:
```
cd services/identity
docker-compose up --build identity-server
```
For now it kind of hangs, but I added explicitly:
```
  println!("Hello");
  return Ok(());
```
in the beginning of the `main` function in the `main.rs` file and it prints `Hello`. That means it successfully copies the proto file, then builds from this proto file and then compiles the sources, so I consider it working.

Reviewers: varun, jimpo

Reviewed By: varun

Subscribers: ashoat, palys-swm, Adrian, atul, yayabosh

Differential Revision: https://phabricator.ashoat.com/D4156