[web] Search for the users on server
Summary:
Another source of the users is our server - there might be some users that aren't in the store, but it should be possible to friend / block them.

Depends on D4267

Test Plan: Render the component and check if all the users that match search text were returned from the server.

Reviewers: def-au1t, atul, benschac

Reviewed By: atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4268