[web] Display user tags
Summary:
Show a list of users that are about to be added. Clicking an `x` button next to their name removes them from the list.

{F80198}

Depends on D4308

Test Plan: Add a couple of users and then remove a couple of them. Every time a name is clicked on the list a tag is added. Every time a tag is closed, the name is added back to the list.

Reviewers: def-au1t, atul

Reviewed By: atul

Subscribers: ashoat, Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4309