[yarn] Restore `fortawesome`-related dependencies to fix icon issue
Summary:
We were experiencing strange behavior with `fortawesome` icons. They were either not appearing correctly (way too large) or not appearing at all.

Linear context:
https://linear.app/comm/issue/ENG-1228/filters-icon-way-too-large-on-web-calendar

I think the `fortawesome/react-fontawesome` dependency needs to be bumped in lockstep with its peer dependencies in order to keep things working.

With this change I just rolled back and fixed version numbers so we're at a state where things look as expected on both `dev` and `prod`.

I'll separately look at bumping these dependencies while preserving the appearance. (Might be a place to try snapshot testing to prevent this sort of regression in the future?)

Test Plan: Tested on `dev` and `prod` and things looked as they used to:

Reviewers: palys-swm, def-au1t, ashoat

Reviewed By: ashoat

Subscribers: Adrian, yayabosh

Differential Revision: https://phab.comm.dev/D4241