[web] Finish styling `ColorSelector` on iOS and Android
Summary:
Completed styling `ColorSelector` on iOS and Android using different devices and breakpoints.

Part of ongoing redesign of `ColorPicker` component. Relevant Linear issue [[ https://linear.app/comm/issue/ENG-1001/redesign-native-colorpicker-component | here ]].

Depends on D3939.

As @atul [[ https://phab.comm.dev/D3939?id=12317#inline-24430 | pointed out ]] in the last diff, any code changes that could break something else shouldn't be committed, but noted in the Summary. This is a change that was made to better encompass the new `ColorSelector`. The height of the `ColorPickerModal` was modified to:
{F72946}

Test Plan:
Looks and works as expected on iOS and Android devices. Note the small and large devices (iPhone SE and iPhone 13 Pro Max, for example):
{F72932}
{F72933}
{F72934}
{F72935}
{F72936}

Reviewers: atul

Reviewed By: atul

Subscribers: varun, ashoat, palys-swm, Adrian

Differential Revision: https://phab.comm.dev/D4254